[IMP] account, web: placeholder in web
In this commit: https://github.com/odoo/odoo/commit/a0f734aeb4771610e00d2d242c4a9cf481b4156b
we added a new widget to add a dynamic placeholder on the field. But we had a
problem when the field already has a widget.

We keep the part needed for the list view in the new widget but the rest is
put in the charField

task:4191835
