+++ addons/account/static/src/components/char_with_placeholder_field/char_with_placeholder_field.xml
@@ -5,9 +5,6 @@
         <xpath expr="//span" position="attributes">
             <attribute name="t-att-class">{'text-muted': !this.props.record.data[props.name]}</attribute>
         </xpath>
-        <xpath expr="//input" position="attributes">
-            <attribute name="t-att-placeholder">placeholder</attribute>
-        </xpath>
     </t>
 
 </templates>
