+++ addons/account/views/account_account_views.xml
@@ -40,7 +40,7 @@
                                         <div>
                                             <field name="placeholder_code" invisible="1"/>
                                             <field name="code" placeholder="e.g. 101000" class="oe_inline"
-                                                   widget="char_with_placeholder_field" options="{'placeholder_field': 'placeholder_code'}"/>
+                                                   options="{'placeholder_field': 'placeholder_code'}"/>
                                         </div>
                                     </div>
                                 </div>
