+++ addons/account/views/account_analytic_distribution_model_views.xml
@@ -27,7 +27,7 @@
                 <data>
                     <xpath expr="//field[@name='partner_category_id']" position="after">
                         <field name="prefix_placeholder" invisible="1"/> <!-- Needed for the char_with_placeholder_field widget -->
-                        <field name="account_prefix" string="Accounts Prefixes" widget="char_with_placeholder_field" options="{'placeholder_field': 'prefix_placeholder'}"/>
+                        <field name="account_prefix" string="Accounts Prefixes" options="{'placeholder_field': 'prefix_placeholder'}"/>
                     </xpath>
                     <xpath expr="//field[@name='company_id']" position="before">
                             <field name="product_id"/>
